requirements.txt
pandas
numpy
requests
beautifulsoup4
trafilatura
sentence-transformers
scikit-learn
umap-learn
hdbscan
keybert
faiss-cpu
PyPDF2
